IGAPP-648: convert disclaimer to typescript
