+++ web-ts/src/routes/__tests__/DisclaimerPage.spec.tsx
@@ -1,12 +1,8 @@
-// @flow
-
-import React from 'react'
-import { shallow } from 'enzyme'
-import moment from 'moment'
-
-import DisclaimerPage from '../DisclaimerPage'
-import { PageModel } from 'api-client'
-
+import React from "react";
+import { shallow } from "enzyme";
+import moment from "moment";
+import DisclaimerPage from "../DisclaimerPage";
+import { PageModel } from "api-client";
 describe('DisclaimerPage', () => {
   const disclaimer = new PageModel({
     path: '/disclaimer',
@@ -14,10 +10,9 @@ describe('DisclaimerPage', () => {
     content: 'this is a test content',
     lastUpdate: moment('2017-11-18T19:30:00.000Z'),
     hash: '2fe6283485a93932'
-  })
-
+  });
   it('should match snapshot', () => {
-    const wrapper = shallow(<DisclaimerPage disclaimer={disclaimer} />)
-    expect(wrapper).toMatchSnapshot()
-  })
-})
+    const wrapper = shallow(<DisclaimerPage disclaimer={disclaimer} />);
+    expect(wrapper).toMatchSnapshot();
+  });
+});
