+++ web-ts/src/routes/__tests__/MainDisclaimerPage.spec.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+import { mount } from "enzyme";
+import MainDisclaimerPage from "../MainDisclaimerPage";
+import { ThemeProvider } from "styled-components";
+import theme from "../../../../modules/theme/constants/theme";
+describe('MainDisclaimerPage', () => {
+  it('should render', () => {
+    mount(<ThemeProvider theme={theme}>
+        <MainDisclaimerPage />
+      </ThemeProvider>);
+  });
+});
